% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{colorizes Text in the otter package}
\usage{
colorize(x, color)
}
\arguments{
\item{x}{Text}

\item{color}{Colour}
}
\value{
Coloured text
}
\description{
colorizes Text in the otter package
}
