% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{open_tutorial}
\alias{open_tutorial}
\title{Open Tutorials in the studyr package}
\usage{
open_tutorial(tutorial_name)
}
\arguments{
\item{tutorial_name}{character string}
}
\value{
opens the tutorial
}
\description{
Open Tutorials in the studyr package
}
\examples{
open_tutorial()
}
