% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter-package.R
\docType{package}
\name{otter-package}
\alias{otter}
\alias{otter-package}
\title{otter: Ott's Utility Functions}
\description{
A collection of personal utility functions, shared in case they are useful to others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ottvahtrik/otter}
  \item \url{https://ottvahtrik.github.io/otter/}
  \item Report bugs at \url{https://github.com/ottvahtrik/otter/issues}
}

}
\author{
\strong{Maintainer}: Ott Vahtrik \email{ott.vahtrik@gmail.com}

}
\keyword{internal}
