% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{show_all_tutorials}
\alias{show_all_tutorials}
\title{Zeige alle Tutorials in Otter}
\usage{
show_all_tutorials()
}
\value{
gibt auf der Konsole aus, welche Tutorials es in Otter gibt und wie deren Kurznamen heißen.
}
\description{
Zeige alle Tutorials in Otter
}
\examples{
library(otter)
show_all_tutorials()
}
