% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{open_tutorial}
\alias{open_tutorial}
\title{Öffnet ein Tutorial in Otter}
\usage{
open_tutorial(tutorial_name = "Grundlagen")
}
\arguments{
\item{tutorial_name}{Gebe den Kurznamen des Tutorials als character string an.
Erhalte die Kurznahmen der Tutorials mit `show_all_tutorials()`.}
}
\value{
öffnet das spezifizierte Tutorial.
Die Standardeinstellung ist auf "Grundlagen" gesetzt.
}
\description{
Öffnet ein Tutorial in Otter
}
\examples{
# Der folgende Code ist als Kommentar geschrieben, und muss entkommentiert
# werden, damit er funktioniert:
#library(otter)
#open_tutorial("Grundlagen")
}
