% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{open_tutorial}
\alias{open_tutorial}
\title{Open Tutorials in the otter package}
\usage{
open_tutorial(tutorial_name = "Grundlagen")
}
\arguments{
\item{tutorial_name}{character string, the short name of the tutorial.
Get the short name of the tutorials using `show_all_tutorials()`.}
}
\value{
opens the specified tutorial. Default value is set to "Grundlagen".
}
\description{
Open Tutorials in the otter package
}
\examples{
library(otter)
#open_tutorial("Grundlagen")
}
