% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{open_tutorial}
\alias{open_tutorial}
\title{import rsconnect
Open Tutorials in the otter package}
\usage{
open_tutorial(tutorial_name)
}
\arguments{
\item{tutorial_name}{character string}
}
\value{
opens the tutorial
}
\description{
import rsconnect
Open Tutorials in the otter package
}
\examples{
library(otter)
#open_tutorial("R_und_RStudio_Grundlagen")
}
