% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otter-package.R
\docType{package}
\name{otter-package}
\alias{otter}
\alias{otter-package}
\title{otter: Ott's Utility Functions}
\description{
A growing collection of personal utility functions. Currently provides tools to parse Rich Text Format (RTF) files and extract their tables into data frames, automatically detecting header rows, merging multi-page tables, and resolving merged cells. Particularly useful for tables produced by SAS or by the 'r2rtf' package, which are commonly used for clinical trial and regulatory reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ottvahtrik/otter}
  \item \url{https://ottvahtrik.github.io/otter/}
  \item Report bugs at \url{https://github.com/ottvahtrik/otter/issues}
}

}
\author{
\strong{Maintainer}: Ott Vahtrik \email{ott.vahtrik@gmail.com}

}
\keyword{internal}
