% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{show_all_tutorials}
\alias{show_all_tutorials}
\title{Show all tutorials in the otter package}
\usage{
show_all_tutorials()
}
\value{
prints out (on the console) all available tutorials in the otter package
}
\description{
Show all tutorials in the otter package
}
\examples{
library(otter)
show_all_tutorials()
}
